Mongodb

+++ src/main/java/de/jvpichowski/rocketsound/messages/base/GpsData.java
@@ -3,12 +3,11 @@ package de.jvpichowski.rocketsound.messages.base;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.telestion.api.message.JsonMessage;
 
-// TODO: float not yet supported by vertx -> changed floats to double (Jan)
 public record GpsData(
 		@JsonProperty int satCount,
 		@JsonProperty boolean fix,
-		@JsonProperty float latitude,
-		@JsonProperty float longitude,
+		@JsonProperty double latitude,
+		@JsonProperty double longitude,
 		@JsonProperty long time) implements JsonMessage {
 	@SuppressWarnings("unused")
 	public GpsData(){
